requirements.txt
openai
python-dotenv
pandas
pydantic
requests
beautifulsoup4
html2text
tiktoken
selenium
readability-lxml
streamlit
streamlit-tags
openpyxl
html2text
webdriver-manager
openpyxl
google-generativeai
tenacity
streamlit-tags